change validity record from to_string function into stringstream input

+++ src/OPTICS/OPTICS.cpp
@@ -495,7 +495,7 @@ void DensityClustering::extractFeatures(const float& radius_eps,
 	ValidityMeasurement vm;
 	vm.computeValue(normOption, ds.dataMatrix, item_cids, object, isPBF);
 	activityList.push_back("Optics Validity measure is: ");
-	timeList.push_back(to_string(vm.f_c));
+	timeList.push_back(to_string(double(vm.f_c)));
 
 	std::cout << "Finishing extracting features!" << std::endl;	
 	IOHandler::printFeature("norm"+to_string(normOption)+"_closest.vtk",
